refactor - rename search tools files
