--- ui/litellm-dashboard/src/components/search_tools/index.tsx
@@ -1,6 +1,0 @@
-export { default as SearchTools } from './search_tools';
-export { SearchToolView } from './search_tool_view';
-export { default as SearchConnectionTest } from './search_connection_test';
-export { SearchToolTester } from './search_tool_tester';
-export * from './types';
-
